refactor(module): refactor starboard into pseudo command and logs
Replace StarboardSettings with unified Starboard and new StarboardLogs
entity and move starboard logic into a pseudo command (ChainEvent-
driven) at src/commands/pseudo/starboard.ts. Remove the old tools
command and separate event files. Also update translator formatting and
apply minor whitespace/formatting tweaks in setting components and
command structure files.
