+++ src/types/structure/command.ts
@@ -391,20 +391,20 @@ export abstract class CustomizableCommand extends BaseCommand {
                                 : row.map((val) => format(setting.format_specifier, val)).join(', ')
                             : `:orange_circle: ${this.t.commands({ caller: 'settings', key: 'execute.not_set', guild_id: BigInt(interaction.guildId!) })}`
                         : this.t.commands({
-                              caller: 'settings',
-                              key: 'execute.view_in_edit_mode',
-                              guild_id: BigInt(interaction.guildId!),
-                          });
+                            caller: 'settings',
+                            key: 'execute.view_in_edit_mode',
+                            guild_id: BigInt(interaction.guildId!),
+                        });
                 } else {
                     value = setting.database_key
                         ? row
                             ? format(setting.format_specifier, row ?? '')
                             : `:orange_circle: ${this.t.commands({ caller: 'settings', key: 'execute.not_set', guild_id: BigInt(interaction.guildId!) })}`
                         : this.t.commands({
-                              caller: 'settings',
-                              key: 'execute.view_in_edit_mode',
-                              guild_id: BigInt(interaction.guildId!),
-                          });
+                            caller: 'settings',
+                            key: 'execute.view_in_edit_mode',
+                            guild_id: BigInt(interaction.guildId!),
+                        });
                 }
                 ui.addFields({
                     name: key,
@@ -457,15 +457,15 @@ export abstract class CustomizableCommand extends BaseCommand {
                             user: user
                                 ? user.tag
                                 : this.t.commands({
-                                      key: 'execute.unknown_user',
-                                      guild_id: BigInt(interaction.guildId!),
-                                  }),
+                                    key: 'execute.unknown_user',
+                                    guild_id: BigInt(interaction.guildId!),
+                                }),
                             date: date
                                 ? date
                                 : this.t.commands({
-                                      key: 'execute.unknown_date',
-                                      guild_id: BigInt(interaction.guildId!),
-                                  }),
+                                    key: 'execute.unknown_date',
+                                    guild_id: BigInt(interaction.guildId!),
+                                }),
                         },
                         guild_id: BigInt(interaction.guildId!),
                     }),
